+++ src/components/CardRegisterForm/index.tsx
@@ -1,22 +1,19 @@
-import {
-  ChangeEvent,
-  Dispatch,
-  FormEventHandler,
-  SetStateAction,
-  useRef,
-  useState,
-} from 'react';
+import { FormEventHandler, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import Card from '../Card';
 import Input, { Focus } from '../Input';
 import Tooltip from '../Tooltip';
 import TooltipButton from '../TooltipButton';
+import useCardRegisterForm from './useCardRegisterForm';
 
-import { isNotAlphabet, isNotNumber } from '../../utils/validation';
 import type { CardInfo } from '../../types/card';
 
 import styles from './cardRegisterForm.module.css';
+
+const today = new Date();
+const currentYear = today.getFullYear() % 100;
+const currentMonth = today.getMonth() + 1;
 
 interface Props {
   registerCard: (card: CardInfo) => void;
@@ -24,37 +21,33 @@ interface Props {
 
 const CardRegisterForm = ({ registerCard }: Props) => {
   const navigate = useNavigate();
-
-  const today = new Date();
-  const currentYear = today.getFullYear() % 100;
-  const currentMonth = today.getMonth() + 1;
-
-  const [cardNumber1, setCardNumber1] = useState('');
-  const [cardNumber2, setCardNumber2] = useState('');
-  const [cardNumber3, setCardNumber3] = useState('');
-  const [cardNumber4, setCardNumber4] = useState('');
-
-  const [expiredMonth, setExpiredMonth] = useState('');
-  const [expiredYear, setExpiredYear] = useState('');
-
-  const [owner, setOwner] = useState('');
-
-  const [cvc, setCvc] = useState('');
-  const [cardPassword1, setCardPassword1] = useState('');
-  const [cardPassword2, setCardPassword2] = useState('');
-
   const inputRefs = Array.from({ length: 10 }).map(() => useRef<Focus>(null));
+  const {
+    cardNumber1,
+    cardNumber2,
+    cardNumber3,
+    cardNumber4,
+    expiredMonth,
+    expiredYear,
+    owner,
+    cvc,
+    cardPassword1,
+    cardPassword2,
 
-  const isValidCardData =
-    cardNumber1.length === 4 &&
-    cardNumber2.length === 4 &&
-    cardNumber3.length === 4 &&
-    cardNumber4.length === 4 &&
-    expiredMonth.length === 2 &&
-    expiredYear.length === 2 &&
-    cvc.length === 3 &&
-    cardPassword1.length === 1 &&
-    cardPassword2.length === 1;
+    setCardNumber1,
+    setCardNumber2,
+    setCardNumber3,
+    setCardNumber4,
+    setExpiredMonth,
+    setExpiredYear,
+    setCvc,
+    setCardPassword1,
+    setCardPassword2,
+
+    isValidCardData,
+    handleNumberChange,
+    handleOwnerChange,
+  } = useCardRegisterForm(inputRefs);
 
   const isValidExpiredDate = (month: number, year: number) => {
     if (month < 1 || month > 12) return false;
@@ -62,40 +55,6 @@ const CardRegisterForm = ({ registerCard }: Props) => {
     if (year === currentYear && month <= currentMonth) return false;
 
     return true;
-  };
-
-  const autoFocusNextInput = (target: HTMLInputElement) => {
-    const { value, maxLength, tabIndex } = target;
-
-    if (tabIndex === inputRefs.length) return;
-
-    if (value.length === maxLength) {
-      inputRefs[tabIndex].current?.focus();
-    }
-  };
-
-  const handleNumberChange = (
-    event: ChangeEvent<HTMLInputElement>,
-    setNumber: Dispatch<SetStateAction<string>>,
-  ) => {
-    const { value } = event.target;
-
-    if (isNotNumber(value)) return;
-
-    setNumber(value);
-
-    autoFocusNextInput(event.target);
-  };
-
-  const handleOwnerChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const { value } = event.target;
-
-    if (value.length === 1 && value === ' ') return;
-    if (isNotAlphabet(value)) return;
-
-    setOwner(value.toUpperCase());
-
-    autoFocusNextInput(event.target);
   };
 
   const handleSubmit: FormEventHandler<HTMLFormElement> = (event) => {
